Moved TRestClient into subfolder
