+++ framework/classes.php
@@ -258,12 +258,6 @@ return [
 'IModule' => 'Prado\IModule',
 'IModuleDependency' => 'Prado\IModuleDependency',
 'ITextWriter' => 'Prado\IO\ITextWriter',
-'TOutputWriter' => 'Prado\IO\TOutputWriter',
-'TStdOutWriter' => 'Prado\IO\TStdOutWriter',
-'TStreamNotificationCallback' => 'Prado\IO\TStreamNotificationCallback',
-'TStreamNotificationParameter' => 'Prado\IO\TStreamNotificationParameter',
-'TTarFileExtractor' => 'Prado\IO\TTarFileExtractor',
-'TTextWriter' => 'Prado\IO\TTextWriter',
 'TCachedHttpClient' => 'Prado\IO\HttpClient\TCachedHttpClient',
 'TCurlHttpClient' => 'Prado\IO\HttpClient\TCurlHttpClient',
 'TFopenHttpClient' => 'Prado\IO\HttpClient\TFopenHttpClient',
@@ -271,7 +265,13 @@ return [
 'THttpClientException' => 'Prado\IO\HttpClient\THttpClientException',
 'THttpClientResponse' => 'Prado\IO\HttpClient\THttpClientResponse',
 'TStreamDownloader' => 'Prado\IO\HttpClient\TStreamDownloader',
-'TRestClient' => 'Prado\IO\TRestClient',
+'TRestClient' => 'Prado\IO\Rest\TRestClient',
+'TOutputWriter' => 'Prado\IO\TOutputWriter',
+'TStdOutWriter' => 'Prado\IO\TStdOutWriter',
+'TStreamNotificationCallback' => 'Prado\IO\TStreamNotificationCallback',
+'TStreamNotificationParameter' => 'Prado\IO\TStreamNotificationParameter',
+'TTarFileExtractor' => 'Prado\IO\TTarFileExtractor',
+'TTextWriter' => 'Prado\IO\TTextWriter',
 'IService' => 'Prado\IService',
 'ISingleton' => 'Prado\ISingleton',
 'IStatePersister' => 'Prado\IStatePersister',
